[Refactor] 폴더명 2차 변경
